feat: updated attendance check logic
- Modified the `insert_attendance_check_records` function to ensure that Attendance Checks are only created for employees with a shift assignment or those whose attendance is tracked by timesheet.
- Removed the `has_no_shift_assignment` field from the Attendance Check doctype.
- Refactored the `insert_attendance_check_records` function to remove a duplicate implementation and improve performance by reducing the number of database queries.
- Updated the test cases to cover the new logic and ensure the bug is fixed.
- Centralized the `create_attendance_check_record` test helper function.
